edit namespace for controllers
